+++ cookbooks/postgresql/files/default/tests/minitest/server_test.rb
@@ -32,12 +32,12 @@ describe 'postgresql::server' do
   it 'can connect to postgresql' do
     require 'pg'
     conn = PG::Connection.new(
-                               :host => '***REMOVED***',
+                               :host => 'localhost',
                                :port => '5432',
                                :password => node['postgresql']['password']['postgres'],
                                :user => "postgres"
                              )
-    assert_match(/***REMOVED***/, conn.host)
+    assert_match(/localhost/, conn.host)
   end
 
 end