+++ wp-content/plugins/gravityforms/includes/addon/class-gf-feed-addon.php
@@ -904,8 +904,8 @@ abstract class GFFeedAddOn extends GFAddOn {
 
 		?>
 
-		<li style="<?php echo $settings_style ?>" id="delay_<?php echo $addon_name; ?>_container">
-			<input type="checkbox" name="paypal_delay_<?php echo $addon_name; ?>" id="paypal_delay_<?php echo $addon_name; ?>" value="1" <?php echo rgar( $feed_meta, "delay_$addon_name" ) ? "checked='checked'" : '' ?> />
+		<div style="<?php echo $settings_style ?>" id="delay_<?php echo $addon_name; ?>_container">
+			<input type="checkbox" name="paypal_delay_<?php echo $addon_name; ?>" id="paypal_delay_<?php echo $addon_name; ?>" value="1" <?php echo rgar( $feed_meta, "delay_$addon_name" ) ? "checked='checked'" : '' ?> class="gaddon-setting gaddon-checkbox" />
 			<label class="inline" for="paypal_delay_<?php echo $addon_name; ?>">
 				<?php
 				if ( rgar( $this->delayed_payment_integration, 'option_label' ) ) {
@@ -915,7 +915,7 @@ abstract class GFFeedAddOn extends GFAddOn {
 				}
 				?>
 			</label>
-		</li>
+		</div>
 
 		<script type="text/javascript">
 			jQuery(document).ready(function ($) {
@@ -1044,7 +1044,11 @@ abstract class GFFeedAddOn extends GFAddOn {
 		return rgar( $paypal_feed['meta'], "delay_{$this->_slug}" ) && $has_payment && ! $is_delayed;
 	}
 
-	public function get_single_submission_feed( $entry, $form = false ) {
+	public function get_single_submission_feed( $entry = false, $form = false ) {
+
+		if( ! $entry && ! $form ) {
+			return false;
+		}
 
 		$feed = false;
 
@@ -1055,7 +1059,12 @@ abstract class GFFeedAddOn extends GFAddOn {
 		} else if( $entry['id'] ) {
 
 			$feeds = $this->get_feeds_by_entry( $entry['id'] );
-			$feed  = empty( $feeds ) ? false : $this->get_feed( $feeds[0] );
+
+			if( empty( $feeds ) ) {
+				$feed = $this->get_single_submission_feed( false, $form );
+			} else {
+				$feed = $this->get_feed( $feeds[0] );
+			}
 
 		} else if ( $form ) {