+++ wp-content/plugins/gravityforms/includes/fields/class-gf-field-multiselect.php
@@ -59,14 +59,17 @@ class GF_Field_MultiSelect extends GF_Field {
 			$size = 7;
 		}
 
-		return sprintf( "<div class='ginput_container'><select multiple='multiple' {$placeholder} size='{$size}' name='input_%d[]' id='%s' {$logic_event} class='%s' $tabindex %s>%s</select></div>", $id, $field_id, $css_class, $disabled_text, GFCommon::get_select_choices( $this, $value ) );
+		return sprintf( "<div class='ginput_container'><select multiple='multiple' {$placeholder} size='{$size}' name='input_%d[]' id='%s' {$logic_event} class='%s' $tabindex %s>%s</select></div>", $id, $field_id, $css_class, $disabled_text, $this->get_choices( $value ) );
+	}
+
+	public function get_choices( $value ){
+		return GFCommon::get_select_choices( $this, $value );
 	}
 
 	public function get_value_entry_list( $value, $entry, $field_id, $columns, $form ) {
 		// add space after comma-delimited values
 		return implode( ', ', explode( ',', $value ) );
 	}
-
 
 	public function get_value_entry_detail( $value, $currency = '', $use_text = false, $format = 'html', $media = 'screen' ) {